move new function to the validator.js file
