+++ repos/fdbt-site/src/pages/api/managePassengerTypes.ts
@@ -13,8 +13,12 @@ import {
     SinglePassengerType,
 } from '../../interfaces';
 import { updateSessionAttribute } from '../../utils/sessions';
-import { getAndValidateNoc, invalidCharactersArePresent, redirectTo, redirectToError } from '../../utils/apiUtils';
-import { checkIntegerIsValid, removeExcessWhiteSpace } from '../../utils/apiUtils/validator';
+import { getAndValidateNoc, redirectTo, redirectToError } from '../../utils/apiUtils';
+import {
+    checkIntegerIsValid,
+    invalidCharactersArePresent,
+    removeExcessWhiteSpace,
+} from '../../utils/apiUtils/validator';
 
 export default async (req: NextApiRequestWithSession, res: NextApiResponse): Promise<void> => {
     try {
